Revert "Refactor `schemamd` and `functionmd` packages to `internal/`"
This reverts commit a1b156e178f678bf37d36248d6bd08ff968e4d49.
