+++ schemamd/testdata/aws_route_table_association.schema.json
