+++ schemamd/testdata/awscc_acmpca_certificate.schema.json
