+++ schemamd/testdata/awscc_logs_log_group.schema.json
