+++ schemamd/write_block_type_description_test.go
@@ -9,8 +9,7 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 	tfjson "github.com/hashicorp/terraform-json"
-
-	"github.com/hashicorp/terraform-plugin-docs/internal/schemamd"
+	"github.com/hashicorp/terraform-plugin-docs/schemamd"
 )
 
 func TestWriteBlockTypeDescription(t *testing.T) {
