+++ schemamd/write_nested_attribute_type_description_test.go
@@ -9,9 +9,8 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 	tfjson "github.com/hashicorp/terraform-json"
+	"github.com/hashicorp/terraform-plugin-docs/schemamd"
 	"github.com/zclconf/go-cty/cty"
-
-	"github.com/hashicorp/terraform-plugin-docs/internal/schemamd"
 )
 
 func TestWriteNestedAttributeTypeDescription(t *testing.T) {
