+++ schemamd/write_type_test.go
@@ -11,7 +11,7 @@ import (
 	"github.com/google/go-cmp/cmp"
 	"github.com/zclconf/go-cty/cty"
 
-	"github.com/hashicorp/terraform-plugin-docs/internal/schemamd"
+	"github.com/hashicorp/terraform-plugin-docs/schemamd"
 )
 
 func TestWriteType(t *testing.T) {
